# dispatcher_microservice/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a dispatcher microservice for Factory Games Organizer api. Api is responsible for creating, updating and authenicating api users, CRUD operations on database associated with the api and provides production calculator service.",
        "title": "Dispatcher microservice",
        "contact": {
            "name": "Marek Banaś",
            "email": "marek.banas004@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1.0"
    },
    "host": "79.175.222.18:40000",
    "basePath": "/",
    "paths": {
...
        "/crud": {
            "put": {
                "security": [
                    {
                        "apiTokenAuth": []
                    }
                ],
                "description": "Updates data in database. Updates the records based on \"id\" field of an element in the array sent in request body. If a record with a particular id does not belong to the user who presented authentication token, then that record is not updated.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CRUD Authorization required"
                ],
                "parameters": [
                    {
                        "description": "Data to be updated in the database",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.JSONDataCrud"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateResponseCrud"
                        }
                    },
                    "400": {
                        "description": "Bad request. One of required parameters is missing or is not of valid format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Authentication error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unexpected serverside error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "apiTokenAuth": []
                    }
                ],
                "description": "Insert data into database. The user to whom the ownership of records is assigned is the user who presented the authentication token.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CRUD Authorization required"
                ],
                "parameters": [
                    {
                        "description": "Data to be inserted into database",
                        "name": "insert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.JSONDataCrud"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.InsertResponseCrud"
                        }
                    },
                    "400": {
                        "description": "Bad request. One of required parameters is missing or is not of valid format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Authentication error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unexpected serverside error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "apiTokenAuth": []
                    }
                ],
                "description": "Deletes data in database. Each table has it's own id list to be deleted. If a record with a particular id does not belong to the user who presented authentication token, then that record is not deleted.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "CRUD Authorization required"
                ],
                "parameters": [
                    {
                        "description": "Data to be deleted in the database",
                        "name": "delete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteInputCrud"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponseCrud"
                        }
                    },
                    "400": {
                        "description": "Bad request. One of required parameters is missing or is not of valid format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Authentication error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unexpected serverside error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/crud/select": {
            "get": {
                "security": [
                    {
                        "apiTokenAuth": []
                    }
                ],
                "description": "Return the records from database specified by id range. Start of range and it's size is specified for each table separately. Size describes number of records to be returned. Ranges include the starting id. Records are only returned for the user that presented authentication token. If start of the range is missing for particular table, then it is assumed to be 1. If size is ommitted, then all records are retreived. If the start of a range is a record belonging to another user, then next record belonging to the user that presented a token is retreived instead.",
                "tags": [
                    "CRUD Authorization required"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of first record to be retreived from machines table",
                        "name": "machines_id_start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of rows to be returned from machines table",
                        "name": "machines_rows",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of first record to be retreived from resources table",
                        "name": "resources_id_start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of rows to be returned from resources table",
                        "name": "resources_rows",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of first record to be retreived from recipes table",
                        "name": "recipes_id_start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of rows to be returned from recipes table",
                        "name": "recipes_rows",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of first record to be retreived from recipes_inputs table",
                        "name": "recipes_inputs_id_start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of rows to be returned from recipes_inputs table",
                        "name": "recipes_inputs_rows",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of first record to be retreived from recipes_outputs table",
                        "name": "recipes_outputs_id_start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of rows to be returned from recipes_outputs table",
                        "name": "recipes_outputs_rows",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of first record to be retreived from machines_recipes table",
                        "name": "machines_recipes_id_start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of rows to be returned from machines_recipes table",
                        "name": "machines_recipes_rows",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.JSONDataCrud"
                        }
                    },
                    "400": {
                        "description": "Bad request. One of required parameters is missing or is not of valid format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Authentication error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unexpected serverside error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/crud/selectbyid": {
            "get": {
                "security": [
                    {
                        "apiTokenAuth": []
                    }
                ],
                "description": "Return the records from database specified by id. Id(s) is specified for each table in the database. If an id parameter for a particular table is omitted, the records are not retreived from that table. Each parameter can be present multiple times, in which case all records from a particular table, with those ids will be retreived and returned in an array. Data is returned for the user that provided authentication token.",
                "tags": [
                    "CRUD Authorization required"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of machines to be retreived from database",
                        "name": "machines_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of resources to be retreived from database",
                        "name": "resources_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of recipes to be retreived from database",
                        "name": "recipes_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of recipes inputs to be retreived from database",
                        "name": "recipes_inputs_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of recipes outputs to be retreived from database",
                        "name": "recipes_outputs_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Id of machines recipes to be retreived from database",
                        "name": "machines_recipes_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.JSONDataCrud"
                        }
                    },
                    "400": {
                        "description": "Bad request. One of required parameters is missing or is not of valid format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Authentication error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unexpected serverside error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/crud/user": {
            "delete": {
                "security": [
                    {
                        "apiTokenAuth": []
                    }
                ],
                "description": "Deletes all data in the database that belongs to user who presented the authentication token.",
                "tags": [
                    "CRUD Authorization required"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponseCrud"
                        }
                    },
                    "400": {
                        "description": "Bad request. One of required parameters is missing or is not of valid format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Authentication error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unexpected serverside error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.DeleteInputCrud": {
            "type": "object",
            "properties": {
                "machinesIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "machinesRecipesIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recipesIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recipesInputsIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recipesOutputsIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "resourcesIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "handler.DeleteResponseCrud": {
            "type": "object",
            "properties": {
                "machinesDeleted": {
                    "type": "integer"
                },
                "machinesRecipesDeleted": {
                    "type": "integer"
                },
                "recipesDeleted": {
                    "type": "integer"
                },
                "recipesInputsDeleted": {
                    "type": "integer"
                },
                "recipesOutputsDeleted": {
                    "type": "integer"
                },
                "resourcesDeleted": {
                    "type": "integer"
                }
            }
        },
...
        "handler.InsertResponseCrud": {
            "type": "object",
            "properties": {
                "machinesInserted": {
                    "type": "integer"
                },
                "machinesRecipesInserted": {
                    "type": "integer"
                },
                "recipesInputsInserted": {
                    "type": "integer"
                },
                "recipesInserted": {
                    "type": "integer"
                },
                "recipesOutputsInserted": {
                    "type": "integer"
                },
                "resourcesInserted": {
                    "type": "integer"
                }
            }
        },
        "handler.JSONDataCrud": {
            "type": "object",
            "properties": {
                "machinesList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.MachineInfo"
                    }
                },
                "machinesRecipesList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.MachinesRecipesInfo"
                    }
                },
                "recipesInputsList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.RecipeInputOutputInfo"
                    }
                },
                "recipesList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.RecipeInfo"
                    }
                },
                "recipesOutputsList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.RecipeInputOutputInfo"
                    }
                },
                "resourcesList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ResourceInfo"
                    }
                }
            }
        },
...
        "handler.MachineInfo": {
            "type": "object",
            "properties": {
                "defaultChoice": {
                    "type": "integer",
                    "format": "int32"
                },
                "id": {
                    "type": "integer"
                },
                "inputsLiquid": {
                    "type": "integer"
                },
                "inputsSolid": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "outputsLiquid": {
                    "type": "integer"
                },
                "outputsSolid": {
                    "type": "integer"
                },
                "powerConsumptionKw": {
                    "type": "integer"
                },
                "speed": {
                    "type": "number",
                    "format": "float32"
                },
                "usersId": {
                    "type": "integer"
                }
            }
        },
        "handler.MachinesRecipesInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "machinesId": {
                    "type": "integer"
                },
                "recipesId": {
                    "type": "integer"
                },
                "usersId": {
                    "type": "integer"
                }
            }
        },
...
        "handler.RecipeInfo": {
            "type": "object",
            "properties": {
                "defaultChoice": {
                    "type": "integer",
                    "format": "int32"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "productionTimeS": {
                    "type": "integer"
                },
                "usersId": {
                    "type": "integer"
                }
            }
        },
        "handler.RecipeInputOutputInfo": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "recipesId": {
                    "type": "integer"
                },
                "resourcesId": {
                    "type": "integer"
                },
                "usersId": {
                    "type": "integer"
                }
            }
        },
        "handler.ResourceInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "liquid": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "resourceUnit": {
                    "type": "string"
                },
                "usersId": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdateResponseCrud": {
            "type": "object",
            "properties": {
                "machinesRecipesUpdated": {
                    "type": "integer"
                },
                "machinesUpdated": {
                    "type": "integer"
                },
                "recipesInputsUpdated": {
                    "type": "integer"
                },
                "recipesOutputsUpdated": {
                    "type": "integer"
                },
                "recipesUpdated": {
                    "type": "integer"
                },
                "resourcesUpdated": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "apiTokenAuth": {
            "type": "apiKey",
            "name": "jwt",
            "in": "query"
        }
    }
}